Allow multiple protocols to be printed at once.
This required modularizing the code better than I had before, so that I
could call it repeatedly on different files.  I also removed the logic
that inserted page breaks after ~50 lines, because those page breaks
become useless if anything is manually written in the notebook or if any
data needs to be inserted.